pyproject.toml
[project]
name = "financial-instrument-qualifier"
version = "0.1.0"
description = ""
authors = [
    {name = "lbatresm"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
